Rename file and correct typo.
